refactor: move bdclient in separate file
