+++ product-service/src/modules/bdService.js
@@ -1,0 +1,23 @@
+import { Client } from 'pg';
+
+const { PG_HOST,
+  PG_PORT,
+  PG_DATABASE,
+  PG_USERNAME,
+  PG_PASSWORD } = process.env;
+
+const dbOptions = {
+  host: PG_HOST,
+  port: PG_PORT,
+  database: PG_DATABASE,
+  user: PG_USERNAME,
+  password: PG_PASSWORD,
+  ssl: {
+    rejectUnauthorized: false
+  },
+  connectionTimeoutMillis: 5000
+}
+
+export const getClient = () => {
+    return new Client(dbOptions);
+}
